Struct saito_rust::crypto::MerkleTree

pub struct MerkleTree<T> {
    pub algorithm: &'static Algorithm,
    // some fields omitted
}

A Merkle tree is a binary tree, with values of type T at the leafs, and where every internal node holds the hash of the concatenation of the hashes of its children nodes.

Fields
algorithm: &'static Algorithm

The hashing algorithm used by this Merkle tree

Implementations

impl<T> MerkleTree<T>

pub fn from_vec(
algorithm: &'static Algorithm,
values: Vec<T, Global>
) -> MerkleTree<T> where
T: Hashable,

Constructs a Merkle Tree from a vector of data blocks. Returns None if values is empty.

pub fn root_hash(&self) -> &Vec<u8, Global>

Returns the root hash of Merkle tree

pub fn height(&self) -> usize

Returns the height of Merkle tree

pub fn count(&self) -> usize

Returns the number of leaves in the Merkle tree

pub fn is_empty(&self) -> bool

Returns whether the Merkle tree is empty or not

pub fn gen_proof(&self, value: T) -> Option<Proof<T>> where
T: Hashable,

Generate an inclusion proof for the given value. Returns None if the given value is not found in the tree.

pub fn gen_nth_proof(&self, n: usize) -> Option<Proof<T>> where
T: Hashable + Clone,

Generate an inclusion proof for the n-th leaf value.

pub fn iter(&self) -> LeavesIterator<'_, T>

Creates an Iterator over the values contained in this Merkle tree.