Struct saito_rust::networking::peer::SaitoPeer

pub struct SaitoPeer { /* fields omitted */ }

A Peer. i.e. another node in the network.

Implementations

impl SaitoPeer

pub fn new(
connection_id: SaitoHash,
host: Option<[u8; 4]>,
port: Option<u16>,
is_connected_or_connecting: bool,
has_completed_handshake: bool,
is_from_peer_list: bool,
wallet_lock: Arc<RwLock<Wallet>>,
mempool_lock: Arc<RwLock<Mempool>>,
blockchain_lock: Arc<RwLock<Blockchain>>
) -> SaitoPeer

pub fn get_is_from_peer_list(&self) -> bool

pub fn set_has_completed_handshake(&mut self, has_completed_handshake: bool)

pub fn get_has_completed_handshake(&self) -> bool

pub fn set_publickey(&mut self, publickey: SaitoPublicKey)

pub fn get_publickey(&self) -> Option<SaitoPublicKey>

pub async fn set_is_connected_or_connecting(
&mut self,
is_connected_or_connecting: bool
)

this setter also will reset the appropriate flags on the peer if the peer becomes disconnected. Once disconnected, we want to also redo the handshake because this will help avoid IP-based connection stealing.

pub fn get_is_connected_or_connecting(&self) -> bool

pub fn get_host(&self) -> Option<[u8; 4]>

pub fn get_port(&self) -> Option<u16>

pub fn get_connection_id(&self) -> SaitoHash

pub fn send_command<'life0, 'life_self, 'async_recursion>(
&'life_self mut self,
command: &'life0 str,
message: Vec<u8>
) -> Pin<Box<dyn Future<Output = Result<APIMessage, APIMessage>> + Send + 'async_recursion>> where
'life0: 'async_recursion,
'life_self: 'async_recursion,

This creates a PeerRequest Future(which sends the APIMessage to the peer), await’s the response, and wraps the returned message into a Result(RESULT__ => Ok(), ERROR___ => Err()).

pub async fn send_command_fire_and_forget(
&mut self,
command: &str,
message: Vec<u8>
)

sends a command to the peer’s socket but doesn’t not await the response

pub async fn send_response_from_str(
&mut self,
message_id: u32,
message_str: &str
)

Helper function for sending basic OK results.

pub async fn send_response(&mut self, message_id: u32, message: Vec<u8>)

Helper function for sending RESULT__.

pub async fn send_error_response_from_str(
&mut self,
message_id: u32,
message_str: &str
)

Helper function for sending basic errors with a string message.

pub async fn send_error_response(&mut self, message_id: u32, message: Vec<u8>)

Helper function for sending errors

pub async fn handle_peer_message(
api_message_orig: APIMessage,
connection_id: SaitoHash
)

handle any APIMessage from the socket. RESULT/ERROR will be matched to it’s COMMAND via the ID and the Future will be polled via it’s waker. Normal commands will be handled by handle_peer_command.