Function saito_rust::peer::send_message_to_socket

pub async fn send_message_to_socket(
    api_message: APIMessage, 
    connection_id: &SaitoHash
)

Sends an APIMessage to a socket connection. Since Outbound and Inbound peers Streams(Sinks) are not unified into a single Trait yet, we must check both dbs to find out which sort of sink this peer is using and send the message through the appopriate stream.