Module saito_rust::slip

Structs

Slip

Enums

SlipType
SlipType is a human-readable indicator of the slip-type, such as in a normal transaction, a VIP-transaction, a rebroadcast transaction or a golden ticket, etc.

Constants

SLIP_SIZE
The size of a serilized slip in bytes.