Crate saito_rust

Welcome to Saito

Saito is a Tier 1 Blockchain Protocol that incentivizes the provision of high-throughput network infrastructure. The network accomplishes with a consensus mechanism that pays the nodes in the peer-to-peer network for the collection and sharing of fees.

Saito-Rust is an implementation of Saito Consensus written in Rust for use by high-throughput routing nodes. It aims to offer the simplest and most scalable implementation of Saito Consensus.

If you need to get in touch with us, please reach out anytime.

Usage

TODO

How to contribute

TODO

Contact

The Saito Team dev@saito.tech

Modules

block

blockchain

blockring

consensus

crypto

mempool

merkle

network

slip

storage

test_utilities

time

transaction

wallet

Type Definitions

Error
Error returned by most functions.

Result
A specialized Result type for operations.